chore(wasm): Share WebGL context creation between renderer and GLCanvasElement
WebGlBrowserRenderer (onscreen Skia surface) and WasmNativeOpenGLWrapper
(GLCanvasElement's offscreen contexts) each carried their own copy of the
window.GL guard, context-attributes block, and WebGL 1.0 fallback. Extract
EmscriptenWebGL.createContext with shared defaults; call sites now only
state their actual differences (antialias for the onscreen surface,
preserveDrawingBuffer for GLCanvasElement readbacks).
